ViT forward pass: gated against PyTorch at cosine 0.999999919
The encoder runs end to end in C and matches open_clip's own activations at
four taps, so a divergence names the stage that caused it rather than just
the final embedding.

Two facts were settled by probing the checkpoint instead of assuming, because
both produce a plausible-but-wrong embedding when guessed:

  Activation is exact GELU, not QuickGELU. open_clip_config.json does not say,
  and BioCLIP-2's LAION-2B lineage only suggests it. The exporter now runs the
  checkpoint's own activation module on fixed inputs and compares against both
  definitions, then records the answer as clip-vit.activation. A file without
  that key is refused rather than defaulted.

  Pooling is CLS (pool_type 'tok'). Also recorded, also refused if unknown.

Verified off the checkpoint at the same time: 24/1024/16/64, eps 1e-5, patch
conv [1024,3,14,14] stride 14 with no bias, proj [1024,768] in [in,out] order
(an nn.Parameter, so NOT transposed like an nn.Linear weight), ls_1/ls_2 are
Identity so there is no LayerScale, and batch_first is True.

Gate against a new F32 export, not the F16 one. The fixtures come from fp32
PyTorch weights, so gating on F16 would fold weight rounding into every
tolerance and leave no budget that means "kernel bug". Tolerances are 10x the
error actually measured (ln_pre 3.8e-6 ... blk_last 4.7e-4), and they hold at
both -O2 and ASAN's -O1, which contract FMAs differently.

What F16 costs is then a measurement rather than a hidden allowance: cosine
0.999999708 vs the fp32 path's 0.999999919, embedding max|abs| 5.2e-3 vs
2.5e-5. That is the baseline the Q8_0/Q6_K/Q4_K numbers get compared against.

  src/dequant.{h,c}  ggml type -> fp32. F32 is a pointer into the mmap and
                     costs nothing; F16 materializes. The seam stage 02 grows.
  src/vit.{h,c}      weights resolved by name (missing tensor = load failure),
                     then the forward pass. The fused qkv weight's three D-row
                     blocks are used in place as Wq/Wk/Wv, no gather.
  tests/test_vit.c   skips when the 1.2 GB model is absent, so a fresh clone
                     still goes green.

Also made STRICT=1 silent again across src/ and tests/ — it had 19 pre-existing
sign-conversion warnings in ops.c, which meant it could not report anything
about new code. Index arithmetic is size_t throughout now; the reference
kernels produce bit-identical output before and after.

+++ tests/test_ops.c
@@ -50,7 +50,7 @@ static void cmp(const char *what, const float *got, const float *want,
         }
     }
     tt_ran++;
-    if (worst_abs > atol && worst_rel > rtol) {
+    if (worst_abs > (double)atol && worst_rel > (double)rtol) {
         tt_fail++;
         fprintf(stderr, "  \033[31mFAIL\033[0m %-16s abs %.3g (at %zu) rel %.3g"
                         "  [atol %.g rtol %.g]\n",
@@ -85,7 +85,7 @@ int main(void)
         snprintf(k, sizeof k, "gemm %s", tags[i]);
         /* K=1024 accumulates ~1e-4 absolute in fp32; that is the kernel being
          * correct, not wrong, which is why both tolerances must be exceeded. */
-        cmp(k, out, c, (size_t)M * N, 2e-4f, 1e-5f);
+        cmp(k, out, c, (size_t)M * (size_t)N, 2e-4f, 1e-5f);
     }
 
     /* ---- linear (with bias, PyTorch weight layout) ---- */
@@ -93,7 +93,7 @@ int main(void)
         int M = U("linear.M"), N = U("linear.N"), K = U("linear.K");
         tc_ref_linear(out, T("linear.x", NULL), T("linear.W", NULL),
                       T("linear.b", NULL), M, N, K);
-        cmp("linear", out, T("linear.y", NULL), (size_t)M * N, 1e-5f, 1e-5f);
+        cmp("linear", out, T("linear.y", NULL), (size_t)M * (size_t)N, 1e-5f, 1e-5f);
     }
 
     /* ---- layernorm (one row offset by 1e4 — the one-pass variance trap) ---- */
@@ -102,7 +102,7 @@ int main(void)
         float eps = gguf_f32_or(&G, "ln.eps", 1e-5f);
         tc_ref_layernorm(out, T("ln.x", NULL), T("ln.gamma", NULL),
                          T("ln.beta", NULL), rows, cols, eps);
-        cmp("layernorm", out, T("ln.out", NULL), (size_t)rows * cols, 1e-4f, 1e-4f);
+        cmp("layernorm", out, T("ln.out", NULL), (size_t)rows * (size_t)cols, 1e-4f, 1e-4f);
     }
 
     /* ---- gelu, both flavours ---- */
@@ -129,10 +129,11 @@ int main(void)
     {
         int rows = U("softmax.rows"), cols = U("softmax.cols");
         tc_ref_softmax_rows(out, T("softmax.x", NULL), rows, cols);
-        cmp("softmax", out, T("softmax.out", NULL), (size_t)rows * cols, 1e-6f, 1e-5f);
+        cmp("softmax", out, T("softmax.out", NULL), (size_t)rows * (size_t)cols, 1e-6f, 1e-5f);
         for (int r = 0; r < rows; r++) {           /* each row must sum to 1 */
             double s = 0.0;
-            for (int c = 0; c < cols; c++) s += out[(size_t)r * cols + c];
+            for (int c = 0; c < cols; c++)
+                s += (double)out[(size_t)r * (size_t)cols + (size_t)c];
             CHECK(fabs(s - 1.0) < 1e-5);
         }
     }
@@ -143,7 +144,7 @@ int main(void)
         float *scratch = (float *)arena_alloc_or_die(&A, (size_t)Tn * sizeof(float), 64);
         tc_ref_mha(out, T("mha.q", NULL), T("mha.k", NULL), T("mha.v", NULL),
                    Tn, H, HD, scratch);
-        cmp("mha", out, T("mha.out", NULL), (size_t)Tn * H * HD, 1e-5f, 1e-5f);
+        cmp("mha", out, T("mha.out", NULL), (size_t)Tn * (size_t)H * (size_t)HD, 1e-5f, 1e-5f);
     }
 
     gguf_close(&G);
